ENH: change formula to compute the shift beteenw pan and xs in phr mode
Change the shift apply in X and Y direction from 3/2 pixels to -3/2 pixels in
bundletoperfectsensor application for phr data. This
patch was provided by Julien Michel as he oberves that it improves the
fusion on PHR data.
I confirm that inverting this offset improve the result on PRIMARY product
samples over Melbourne.

It is still not clear for me if we've got now the "right" formula.