link libjpeg against DLL version of mscvrt library
makefile.vc builds libjpeg static with /MT flags that
link with static mscvt. This libjpeg when linking against
gdal throws link error as gdal is linked against MSVCRT
dll version so as all other dependencies in superbuild.
This makes external libjpeg currently unusabe with gdal
and end upwith two libjpeg in the build. This commit
build libjpeg as cmake project which automatically add /MD
depending on configuration and finally to have a single libjpeg
used in GDAL and libtiff

http://dash.orfeo-toolbox.org/testDetails.php?test=28153014&build=170814