include numpy.i from github/numpy for numpy support in ApplicationEngine
numpy.i defines swig typemaps for numpy array. These are used in
otbApplication.i to feed a numpy in and out of application.
This file is supposed to be copied in any application it uses as it is
not distributed with SWIG. License of numpy.i is listed on file header
and is NOT part of OTB and its license.

Furthermore, Inclusion of numpy.i in other projects has been
discussed with numpy developers on their github tracker.
See issue #5690 -https://github.com/numpy/numpy/issues/5690

Current version of this file is taken from numpy/tools/swig.
github commit: b89bcd05f10001b048ffd12bdaa550bcd0620fa0

Signed-off-by: Rashad Kanavath <rashad.kanavath@c-s.fr>