SuperBuild: rename patches with a specific 'unstandard' naming convention.
This style of naming can later be used to automate patching for
dependencies. For instance take the case of gdal.
the name of current patch is gdal-fix-rpath-for-macx.diff
This will imply that project is GDAL patch description is
second part 'fix-rpath' and for-macx.diff says that
it is only for macx.
Later (in future) cmake can use if(APPLE), if(WIN32),
if(UNIX and NOT APPLE) and see if the name of diff file
matches for-macx, for-win32, for-linux and apply it accordingly.
The project name prefix seems redundant at first look.
But later this can pave way to have all .diff patches in a
single directory under superbuild rather than each one in sub-directory
named 'project'. The former can help in case of libraries like FFTW
where it builds two projects FFTWF and FFTWD.