ENH: Modify internal implementation of InputImageListParameter to act as a vector of InputImageParameter
External API remains unchanged, but we can now benefit from all the services of InputImageParameter, including boiler-plate template/macro code to cast ImageBase to any type of image.

Also, this avoids a lot of code duplication and could be used in the future to also avoid duplication in QT widget part.