BUG: Add a workaround to check the model name before serialization.
This workaround is added to avoid a crash in application when build
is made with RelWithDebInfo which is caused by a
boost::archive_exception impossible to be catch.
This workaround break the compatibility with shark model direct
read/write.
It is still possible to read old RF model from OTB.