ImageFileReader: Fix misleading error handling for directories
GetGdalReadImageFileName does not set its second argument when its first
argument is a non empty directory and not a supported format (OTB
supports opening directories in some cases). But in that case it does
return false, which was not handled before this commit. This resulted in
a misleading error message and a corrupted m_FileName.