WRG: hide msvc D9025 warning on dashboard
This bitching warning seems too annoying on dashboard
superbuild-contrib build. total number of warnings seems
300. This magic number is the maximum number of dashboard
warnings in OTB. superbuild usually set max warnings to zero
in their CTestCustom.cmake.in. But here the issue is when
doing superbuild-contrib we switch binary directory to OTB
and run configure, build, install from it and not from
SuperBuild binary directory.

The case about D9025 warning on msvc is actually a misconfig
in the list of arguments passed. The related cause is from
Modules/ThirdParty/6S/src/CMakeLists.txt where we force /Ob1
flag for msvc > 1800. the change is decorated with
enough inline docs. As this is a commandline warning, it cannot
be silenced using #pragma and moreover it is important to show
it to user. But for dashboard case, it is too annoying for
developers