ENH: upgrade makeself script
the current version is makeself has some issue on running linux
particulary on Ubuntu 14.04. I suspect this is the issue on
test passed (but actually failed) only on linux.

Linux (passed with no output messages):
https://dash.orfeo-toolbox.org/testDetails.php?test=59547315&build=287532
OSX (passed with proper messages):
https://dash.orfeo-toolbox.org/testDetails.php?test=60023362&build=287444

lurking on makeself's bug tracker, it could be a umask issue
https://github.com/megastep/makeself/issues/58
https://github.com/megastep/makeself/commit/842dedfe1a1477237c5132dc3615fcf50dddcc65
https://github.com/megastep/makeself/commit/e75285d90df44fb7bd80a4f727442e784f6e73c5
I am not sure if this is the exact problem and upgrade will fix
issue for us. Because we are not able to reproduce this bug on same
machine executed under same constraints.
Only change is that one is running via crontab and other directly from
terminal.