BUG: fix sse test for msvc, gnu, clang
otbsiftfast was enabling /fp:fast option with msvc along with SSE.
This is not normal as MSDN docs doesn't mention anything about /fp:fast
when using /arch:SSE.

I think it has a problematic behaviour with otb's dashboard
testing. Many --compare-image tests we do are with zero tolerance.
So it has to go.