BUG: error generating qgis descriptors
below commit uses POST_BUILD and PRE_BUILD to generate a fresh copy of
descriptor files for all OTB applications.
https://gitlab.orfeo-toolbox.org/orfeotoolbox/otb/commit/b963a17746d854656e83c9774845c850d117a37d

This has broken dashboard due to a dependency issue.
https://dash.orfeo-toolbox.org/buildSummary.php?buildid=318175

otbQgisDescriptor is not dependent on any otb applications and it is
TRUE. But it has a run-time dependency to all otb applications when we
auto-generating all depscriptors. This situation forces us to add a
dependency so that POST, PRE building targets works. These are only
executed if otbQgisDescriptor target is rebuilt hence, it is not run
everytime with make.

Revert of the initial commit[2] is to get source tree in to a stage which
was reviewed. The first version of post,pre build was pushed directly to
release rather than via MR. Current state in release-6.6 will result in
broken dashboard.

[2] https://gitlab.orfeo-toolbox.org/orfeotoolbox/otb/commit/3f7b6d401bafb1975051299e5d2a9f337dcb1bd8