ENH: install share dir share/OTB-X.Y -> share/otb
qgis descriptors are copied to share/otb/description.
This ends up with share/otb/ and share/OTB-X.Y
It is better to have a single directory for and it should be
share/otb and not a versioned. Changes included in packaging
script show non-versioned 'otb' directory in share is easy to
manage than a verisoned one. In any case, users cannot keep
multiple versions of OTB under same prefix.
This was breifly discussed in issue :
https://gitlab.orfeo-toolbox.org/orfeotoolbox/otb/issues/1557