BUG: fix issue !1639 Update UserValue flag when a parameter is modified in QtWidgetParameterList widget
The QtWidgetParameterList widget should signal that an input parameter as been
modify by the user.Then, it set the UserValue flag from the corresponding parameter to
True.
It was spotted as the UserValue flag is used by the ParseApplication method used
to list parameter modified by the user in the GUI which is used to build the
corresponding command-line from the application GUI (displayed in the Log tab).