ENH: rename files with txx extension to hxx
OTB followed since the beginning the ITK convention and use .txx extension for all
template classes. Nevertheless, some development tools do not recognize .txx
file extension. Other tool like GitHub can't do in-browser syntax highlighting for
txx files I think.

The root problem is the use of the txx which should be changed to hxx (or hpp).

In 2011, after an in-depth discussion near April 20, 2011 on the
Insight-Developers mailing list, ITK rename all txx files to hxx (and event
prevent the push of .txx files with a pre-commit hook). It happens is major release v4.

You can find some arguments in the discussion about the change and also in other
projects related to ITK which applied the same modification, see for instance VXL:

https://github.com/vxl/vxl/issues/209

This commit apply now the same modification for OTB.

I understand that it will change some habit for developers and don't bring new
features but I think that in general it is better to stay align with ITK guidelines.

In my opinion, it always facilitate the use of OTB and ITK together if we share
when we can the same code architecture, directory organization, naming
conventions...