ENH: Allow to compute modulus and phase from 2 bands scalar image
The support of complex images in otb applications as improved in version
6.6. There is generic mechanism based on teh ClampImageFilter which allow to
go from/to complex or multi-bands scalar images.

The only change is to read the input image as ComplexFloatImageType which
support both complex image (single band) or image with 2 bands which corresponds
to the real and imaginary parts.

The documentation of the application has been updated to reflect this
modification.