COMP: fix dynamic multithreading to use ITK 5.0
ITK 5.0 uses a different threading model compared to ITK
4.x. DynamicThreadedGenerateData() is the newer variant without
threadId, and is the preferred signature, which is called by
default. This variant can split the requested region into different
number of pieces depending on current multi-processing load, which
allows better load balancing. The non-dynamic (also known as classic)
ThreadedGenerateData() signature has threadId, and number of pieces to
be split into is known in advance. It is activated by calling
this->DynamicMultiThreadingOff(); in derived class
constructor. OTB_DISABLE_DYNAMIC_MT is a simple macro defined in
otbMacro.h that insert this function call *IF* itk version used is 5.0 or above.

Sample run-time exception:
 2018-06-04 15:30:09 (DEBUG): Caught itk::ExceptionObject during application execution:
 2018-06-04 15:30:09 (DEBUG): /usr/src/ports/itk/Modules/Core/Common/src/itkPoolMultiThreader.cxx:202:
 itk::ERROR: PoolMultiThreader(0xabad1dea): Exception occurred during SingleMethodExecute
 /home/rashad/local/include/ITK-5.0/itkImageSource.hxx:276:
 itk::ERROR: ShiftScaleImageFilter(0xabad1deaff): Subclass should override this method!!!
 If old behavior is desired invoke this->DynamicMultiThreadingOff(); before Update() is called. The best place is in class constructor.