Wrapping: raise swig minimum version to 3.0
SWIG 3.0 or higher is required with ITK 5.0. This is because older
version of swig cannot handle c++11 keywords. swig simply send out a
syntax error:
/home/rashad/local/include/ITK-5.0/itkFloatTypes.h:30: Error: Syntax
error in input(1)

If we make minimum version to be 3.0, atleast we get a clear error on
configure which is much better than cryptic error messages