ENH: fix LexicographicalCompare in ITK 4, 5(alpha1, alpha2), 5(master)
iktLexicographicCompare.h is new in ITK 5 which was using a templated
class until ITK 5 (alpha2) release. It was made "better" by not using
template parameter for this class. Initial approach was to use ifdef on
ITK_VERSION_MAJOR but that is going to end up in more problems since
different itk version with same major have differences. So We copy old
class def from ITK4.x and use it directly in one class where it is used