BUG: min required swig version is 3.0.10
on ubuntu 14.04 LTS system package has swig version 3.0.2 but this give
a warning when parsing itkBase.i

"Warning 341: The 'using' keyword in type aliasing is not fully
supported yet."

I had tested swig 3.0.10 (latest is 3.0.12) on windows and linux which
does not have this warning and works fine for python and java. Even
though this is a "warning" when parsing .i subsequent compilation fails
with swig 3.0.x.