ENH: upgrade cmake min version to 3.10.2
ITK beta1 has cmake min version set to 3.10.2.
https://github.com/InsightSoftwareConsortium/ITK/blob/master/CMakeLists.txt

So we all have to build superbuild with cmake 3.10.2. Alternative is to
ask for a cmake >= 3.10.2 when building ITK. I think this is better path
than using an if else. This way all libraries can use same cool features
like in ITK project.