ENH: use GIT_TAG master for all remote modules
Changes in OTB, platform, compilers, upgrade of dependencies might need
a change in remote module code. This code is maintained (mostly) by
people not invovled in everyday activity of OTB. Due to delay in make
changes dashboard of OTB will be broken. It keeps happening again.

Current workflow:
1. create issue on remote module repo
2. review/propose a merge request on that project
3. test this remote module (manually) on all platforms
4. create a new mr to update GIT_TAG

New workflow:
1. use "master" for GIT_TAG
2. allow remote module devs to continue pushing fixes
3. GIT_TAG will be updated during release process after creation of
release-X.Y branch

Benenfits:
* developers of OTB can test latest version of a remote module
when using develop
* users can get access to latest version of a remote module
when using develop
* developers of remote module can test latest changes with latest OTB
on all platforms (using cdash)
* reduce delay and avoid updating GIT_TAG with each module in between
  releases.

when a build fails because changes in OTB are not reflected in remote module. *could we disable this module.
remote module are nice. but it is now fixing their code and then a mr to update tag.