PERF: Localize reduce computation
The reduce computation done in `DifferenceImageFilter` was done on some
`m_ThreadAccumulatedCounter[threadId]` member. This has two drawbacks:
- first the compiler cannot know whether we are the only one function to work
  on that value, it cannot isolate it in a register nor anything similar
- moreover _false sharing_ happens and prevent further optimizations.

As a consequence the counters are now local variables in the
`ThreadedGenerateData()` function, and they are commited into the shared member
variables at the end of the execution.

Note: this anti-pattern appears in other places like for instance
`StreamingCompareImageFilter`.