The CI must warn the team of new contributors who haven't signed the CLA
To legally secure the OTB project, the new contributors must sign a
contributor license agreement (CLA). The script added to the IC
pipeline checks that the contributors who appear in the Git history
have signed this agreement. This list of contributors is compared with
an external reference list. In case of a difference, the job fails and
lists the unknown identities.