Sort application names to get a sorted algs.txt
Application descriptions are written for each application in
OTB_APPLICATIONS_NAME_LIST. This list is unsorted, which results in
descriptions written in random order in algs.txt.

In order to get a deterministic output in algs.txt and thus a more
reproducible build, make sure to iterate over a sorted list of application
names.