ENH: Refactor the IMI and the factory
The IMI don't need the m_Imd attribut. This brings extra copies.
Instead, the parse method takes the imd as a reference and
completes it.
The IMI depends heavily on the MetadataSupplierInterface (mds)
provided to the parse methode. Let's remove it from the
parse method, and create an attribut instead.