PERF: Move DEMHangling wrapper to TLS
The previous design was using a single wrapper around GDALDataSet. But all
accesses were thread locked. More than half of the execution time was spent in
mutex locking and unlocking.

The new design uses a GDALDataSet per thread. All GDALDataSet point to the same
file.
As per
https://trac.osgeo.org/gdal/wiki/FAQMiscellaneous#IstheGDALlibrarythread-safe,
this new design is sound with GDAL 2.0.

In a direct DEM to XYZ transformation, computation time goes from ~8h to ~1h15.

The current change is half proof of concept, half quick and dirty workaround.
In a next iteration, datasets should not be exclusive to one thread but kept in
a datapool for reuse between OTB streams.