PERF: Factorize coordinate transformation for GWS84
Searching for whether we are in GWS84 can be costly (10% of the time in a DEM
to XYZ application).
=> The `OGRCoordinateTransformation` object creation is factorized when the TLS
DEMHandler is created -- as there is no reason for changes once the
`GDALDataset` is instanciated.

This also indirectly reduce the number mutex locked through
`OGRSpatialReference::GetWGS84SRS()` (~9% of the GDAL mutex locked in the
scenario).