WIP: Stabilize DEMHandler refactoring
It seems it was not possible to guarantee the destruction of static
thread_local Meyer singletons before the destruction of the global Meyer
singleton.

The code has been refactored to use `sahred_ptr` to automatically countthe
number of owners behind a `DEMHandlerTLS`: at max: a thread + the global
singleton.

Also, several datasets are created for accessing the same VRT file.

TODO: check whether the VRT shouldn't be created once from the main thread
(i.e. from the global `DEMHandler`), but the VRT should then be opened once per
`DEMHandlerTLS`, but then beware of the concurrent changes on opened file.