BUILD: correctly transfer cmake GDAL args to disable spatialite
Cmake arguments were not transfered due to options not used as cmake list. Thus
when spatialite was installed, gdal automatically detects it.
As spatialite can use a system libproj which can be different to the one we use
it creates a gdal lib linked with the two libproj. This lead to lib symbols errors
during runtime.